Kenzie/Sub-MAF-07: Created a listener for languages
